命名規則をCreateCharacterに統一。
